moved ReadRowsIterator into new file
