+++ google/cloud/bigtable/iterators.py
@@ -1,0 +1,108 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from __future__ import annotations
+
+from typing import (
+    cast,
+    AsyncIterable,
+)
+import asyncio
+import time
+import sys
+
+from google.cloud.bigtable._row_merger import _RowMerger
+from google.cloud.bigtable_v2.types import RequestStats
+from google.api_core import exceptions as core_exceptions
+from google.cloud.bigtable.exceptions import RetryExceptionGroup
+from google.cloud.bigtable.exceptions import IdleTimeout
+from google.cloud.bigtable.row import Row
+
+
+class ReadRowsIterator(AsyncIterable[Row]):
+    """
+    User-facing async generator for streaming read_rows responses
+    """
+
+    def __init__(self, merger: _RowMerger):
+        self._merger_or_error: _RowMerger | Exception = merger
+        self.request_stats: RequestStats | None = None
+        self.last_interaction_time = time.time()
+        self._idle_timeout_task: asyncio.Task[None] | None = None
+
+    async def _start_idle_timer(self, idle_timeout: float):
+        self.last_interaction_time = time.time()
+        if self._idle_timeout_task is not None:
+            self._idle_timeout_task.cancel()
+        self._idle_timeout_task = asyncio.create_task(
+            self._idle_timeout_coroutine(idle_timeout)
+        )
+        if sys.version_info >= (3, 8):
+            self._idle_timeout_task.name = "ReadRowsIterator._idle_timeout"
+
+    def active(self):
+        return isinstance(self._merger_or_error, _RowMerger)
+
+    async def _idle_timeout_coroutine(self, idle_timeout: float):
+        while self.active():
+            next_timeout = self.last_interaction_time + idle_timeout
+            await asyncio.sleep(next_timeout - time.time())
+            if (
+                self.last_interaction_time + idle_timeout < time.time()
+                and self.active()
+            ):
+                # idle timeout has expired
+                await self._finish_with_error(IdleTimeout("idle timeout expired"))
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self) -> Row:
+        if isinstance(self._merger_or_error, Exception):
+            raise self._merger_or_error
+        else:
+            merger = cast(_RowMerger, self._merger_or_error)
+            try:
+                self.last_interaction_time = time.time()
+                next_item = await merger.__anext__()
+                if isinstance(next_item, RequestStats):
+                    self.request_stats = next_item
+                    return await self.__anext__()
+                else:
+                    return next_item
+            except core_exceptions.RetryError:
+                # raised by AsyncRetry after operation deadline exceeded
+                new_exc = core_exceptions.DeadlineExceeded(
+                    f"operation_timeout of {merger.operation_timeout:0.1f}s exceeded"
+                )
+                source_exc = None
+                if merger.errors:
+                    source_exc = RetryExceptionGroup(
+                        f"{len(merger.errors)} failed attempts", merger.errors
+                    )
+                new_exc.__cause__ = source_exc
+                await self._finish_with_error(new_exc)
+                raise new_exc from source_exc
+            except Exception as e:
+                await self._finish_with_error(e)
+                raise e
+
+    async def _finish_with_error(self, e: Exception):
+        if isinstance(self._merger_or_error, _RowMerger):
+            await self._merger_or_error.aclose()
+            del self._merger_or_error
+            self._merger_or_error = e
+        if self._idle_timeout_task is not None:
+            self._idle_timeout_task.cancel()
+            self._idle_timeout_task = None
